Fix: debug joystick + warning C4244 residui
1. DEBUG JOYSTICK (Game.cpp)
   Aggiunto output diagnostico per capire perché il player non si muove:
   - initJoystickSystem(): stampa i valori di tutti gli 8 assi (X, Y, Z,
     R, U, V, PovX, PovY) per ogni joystick collegato. Cosi vediamo
     quali assi sono attivi sul gamepad.
   - In update() (STATE_PLAYING): stampa i valori degli assi P1 quando
     sono > 10 (movimento rilevato), ogni ~0.5s per non spam.
     Mostra: jid, axisX config, axisY config, valori x/y attuali.

   L'utente dovra' guardare l'output console e dirmi:
   a) Quali assi cambiano valore quando muove l'analogico
   b) Se 'JOY P1' appare quando muove il stick
   Questo ci dira' se il problema e':
   - assi sbagliati (es. gamepad usa Z/R invece di X/Y)
   - getAxisPosition restituisce 0 (sf::Joystick::update non funziona)
   - deadzone troppo alta (30)
   - config.ini con valori sbagliati

2. FIX WARNING C4244 RESIDUI
   a) Utils.h + Utils.cpp: drawText/drawTextCentered/drawTextOutlined/
      drawTextCenteredOutlined - cambiati parametri cx/y da float a int.
      Questo elimina ~15 warning C4244 (int->float) nelle chiamate con
      WINDOW_WIDTH/2 ecc. Le funzioni castano internamente a float.
   b) Game.cpp: 3 chiamate con argomenti float castate a (int):
      - drawTextCenteredOutlined(... centerY + wheelRadiusY + 80.f)
        -> (int)(centerY + wheelRadiusY + 80.f)
      - drawTextCentered('ENTRA', dx, dy - 30.f + bobY)
        -> (int)dx, (int)(dy - 30.f + bobY)
   c) Player.cpp: drawTextCentered(weapon name, px, pos.y - 60)
      -> (int)px, (int)(pos.y - 60)
   d) Enemy.cpp L902/904: setPosition(... py - 2 + i*3)
      -> py - 2.f + (float)(i*3) (gia' fixato nel commit precedente)
   e) AudioManager.cpp L730: int = double (samplesPerSixteenth)
      -> (int)(SR * sixteenthDur) (gia' fixato)

Compila senza errori (g++ -std=c++17 -fsyntax-only -Wall -Wextra).

+++ src/Player.cpp
@@ -336,7 +336,7 @@ void Player::render(sf::RenderTarget& target) {
     // py non piu' usato dopo la refactoring del fallback procedurale
 
     // Etichetta arma sopra la testa
-    drawTextCentered(target, currentWeapon.getName(), px, pos.y - 60, 2, sf::Color(255, 255, 0));
+    drawTextCentered(target, currentWeapon.getName(), (int)px, (int)(pos.y - 60), 2, sf::Color(255, 255, 0));
 
     // Tentativo di rendering con sprite.
     // Animazioni: attack (se shootAnimTimer>0) > walk > idle.
